+++ CHANGELOG.md
@@ -4,6 +4,103 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [1.120.1004] - 2024-06-11
+
+### Added
+
+- `viewDependent` parameter has been added to the **openAsync()** and **open()** methods of the **FragmentCL** class. If this parameter is set to true, dialog or popover is added as a dependent to the source view using the **addDependent()** method of the MVC View class.
+- **closeAndDestroy()** method has been added to the **FragmentCL** class. With this method, it is no longer necessary to call **close()** and then **destroyFragmentContent()**. This method does both.
+
+## [1.124.1002] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
+## [1.124.999002] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
+## [1.123.2002] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
+## [1.123.1003] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
+## [1.120.15002] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
+## [1.120.14002] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
+## [1.120.13002] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
+## [1.120.12002] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
+## [1.120.11002] - 2024-06-11
+
+### Added
+
+- `filterCaseSensitive?` parameter added to the constructor of the ValueHelpCL to change the default value of case-sensitive search in the filterbar.
+
+### Fixed
+
+- ValidationLogicCL bug fix for falsy values such as false, 0, "" in the `value1` and `value2` properties. 
+
 ## [1.120.1003] - 2024-06-11
 
 ### Added
@@ -310,6 +407,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Initial productive release.
 
+[1.120.1004]: https://github.com/hasanciftci26/ui5-antares/compare/1.120.1003...1.120.1004?diff=split
+[1.124.1002]: https://github.com/hasanciftci26/ui5-antares/compare/1.124.1001...1.124.1002?diff=split
+[1.124.999002]: https://github.com/hasanciftci26/ui5-antares/compare/1.124.999001...1.124.999002?diff=split
+[1.123.2002]: https://github.com/hasanciftci26/ui5-antares/compare/1.123.2001...1.123.2002?diff=split
+[1.123.1003]: https://github.com/hasanciftci26/ui5-antares/compare/1.123.1002...1.123.1003?diff=split
+[1.120.15002]: https://github.com/hasanciftci26/ui5-antares/compare/1.120.15001...1.120.15002?diff=split
+[1.120.14002]: https://github.com/hasanciftci26/ui5-antares/compare/1.120.14001...1.120.14002?diff=split
+[1.120.13002]: https://github.com/hasanciftci26/ui5-antares/compare/1.120.13001...1.120.13002?diff=split
+[1.120.12002]: https://github.com/hasanciftci26/ui5-antares/compare/1.120.12001...1.120.12002?diff=split
+[1.120.11002]: https://github.com/hasanciftci26/ui5-antares/compare/1.120.11001...1.120.11002?diff=split
 [1.120.1003]: https://github.com/hasanciftci26/ui5-antares/compare/1.120.1002...1.120.1003?diff=split
 [1.108.32002]: https://github.com/hasanciftci26/ui5-antares/compare/1.108.32001...1.108.32002?diff=split
 [1.108.31002]: https://github.com/hasanciftci26/ui5-antares/compare/1.108.31001...1.108.31002?diff=split